Refactor again FarPatchTableFactory. - rename "Regular end cap" to "BSplineBasis end cap" - revert templating and add EndCapType into PatchTablesFactory::Options. - make EndCapFactories internal in PatchTablesFactory. - move end cap stencils into PatchTables, keep them relative to the max level. - add a utility StencilTablesFactory::AppendEndCapStencilTables to splice and factorize endcap stencil tables.

+++ examples/vtrViewer/gl_mesh.cpp
@@ -496,8 +496,8 @@ GLMesh::InitializeFVar(Options options, TopologyRefiner const & refiner,
 
             for (int edge=0; edge<nedgesperpatch; ++edge) {
 
-                eao.push_back(eao.size());
-                eao.push_back(eao.size());
+                eao.push_back((int)eao.size());
+                eao.push_back((int)eao.size());
 
                 int v0 = offset + basisedges[edge*2],
                     v1 = offset + basisedges[edge*2+1];
